// admin/src/models/product.ts
import { Category } from "./category"

export interface Product {
  id: string;
  title: string;
  description: string;
  image: string;
  type: string;
  price: number;
  unit: string;
  salePrice: number;
  discountInPercent: number;
  category: Category;
  gallery: Gallery[];
}

export type Gallery = {
  url: string;
}

export interface ProductCart extends Product {
  quantity: number;
}